separate ModalDialog source into it's own file.  no export from repo
